Carry over arguments from failed actions
When the bot tries to run some function with some arguments and an error is raised, the
submitted arguments are lost and need tobe resubmitted.
Now, we can carry over the arguments that were valid (assumes that each arguments are
checked in the order that they are submitted)
The BadInputError now carries over the good arguments.
The brain.process now processese keywords 'status' for current function in progress, and
'undo' to remove the last argument submitted